Assignment 3.4
Problem Statement

A.	Implement user defined functions within apply function using the mtcars data set and produce column wise summary statistics using apply function and mtcars dataset.
Answer	: 
apply() Function
The apply() function is most often used to apply a function to
the rows or columns (margins) of matrices or data frames.
However, it can be used with general arrays, for example, to
take the average of an array of matrices. Using apply() is not
faster than using a loop function, but it is highly compact and
can be written in one line.
The syntax for apply() is as follows where
• x is the matrix, data frame or array
• MARGIN is a vector giving the subscripts which the
function will be applied over. E.g., for a matrix 1 indicates
rows, 2 indicates columns, c(1, 2) indicates rows and
columns.
• FUN is the function to be applied
• ... is for any other arguments to be passed to the function
# syntax of apply function
apply(x, MARGIN, FUN, ...)
# show first few rows of mtcars
head(mtcars)
## 		mpg cyl disp hp drat wt qsec vs am gear carb
## Mazda RX4   	21.0 6 160 110 3.90 2.620 16.46 0 1 4 4
## Mazda RX4 Wag 	      21.0 6 160 110 3.90 2.875 17.02 0 1 4 4
## Datsun 710         22.8 4 108 93 3.85 2.320 18.61 1 1 4 1
## Hornet 4 Drive     21.4 6 258 110 3.08 3.215 19.44 1 0 3 1
## Hornet Sportabout 18.7 8 360 175 3.15 3.440 17.02 0 0 3 2
## Valiant            18.1 6 225 105 2.76 3.460 20.22 1 0 3 1 
B.	write a program to extract the names of the list.

Answer		:  

data <- list(item1 = 1:4,
             item2 = rnorm(10),
             item3 = rnorm(20, 1),
             item4 = rnorm(100, 5)